+++ src/__tests__/buildConfig-test.js
@@ -236,6 +236,9 @@ test("lookup configuration from file's path", t => {
       rules: {
         "block-no-empty": true,
         "color-no-invalid-hex": true,
+        indentation: 2,
+        "no-eol-whitespace": true,
+        "no-missing-end-of-source-newline": true,
       },
     })
     t.end()