+++ src/__tests__/postcssPlugin-test.js
@@ -48,7 +48,7 @@ test("lookup configuration file by css file's path", t => {
     from: path.join(__dirname, "fixtures/lookupConfig/a.css"),
   }
   t.plan(2)
-  postcssPlugin.process("a {color: #zzzzzz}", config).then(result => {
+  postcssPlugin.process("a {color: #zzzzzz}\n", config).then(result => {
     const warnings = result.warnings()
     t.equal(warnings.length, 1)
     t.ok(warnings[0].text.indexOf("color-no-invalid-hex") !== -1)