+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Head
 
+-   Fixed: When look for configuration file, starting from source file path, not `process.cwd()`.
+-   Fixed: If no `configFile` is passed, and the `config` object passed uses relative paths for `extends` or `plugins`, you are going to have to pass a `configBasedir`. If not, you do not need this.
+-   Fixed: When config.ignoreFiles has a absolute path non glob (eg: `!/proj/*`). No wrong convert.
+-   Added: if `resolveFrom` fail and `lookup` is id, try resolve from `process.cwd()`
+-   Added: extend configs can be a function now.
 -   Fixed: `block-no-single-line` now ignores empty blocks.
 -   Fixed: `--ignore-path` and `--report-needless-disables` no longer fails when used together.
 -   Fixed: the `indentation` rule now correctly handles `_` hacks on property names.