+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 -   Fixed: `block-no-single-line` now ignores empty blocks.
 -   Fixed: `--ignore-path` and `--report-needless-disables` no longer fails when used together.
 -   Fixed: the `indentation` rule now correctly handles `_` hacks on property names.
+-   Fixed: the `property-no-unknown` rule now ignores SCSS nested properties.
 
 # 7.2.0